requirements.txt
netaddr
djangorestframework
